feat: make mark apis

+++ backend/src/services/student/v1/delete.student.service.ts
@@ -1,12 +1,20 @@
 import { RequestHandler } from "express";
 import { prisma } from "../../../db";
 import { HttpStatus } from "../../../enums";
+import { ApiError } from "../../../utils/ApiError";
 import { ApiResponse } from "../../../utils/ApiResponse";
 import { asyncHandler } from "../../../utils/asyncHandler";
 import { getStudentValidator } from "./validators/get.student.validator";
 
 const deleteStudent: RequestHandler = asyncHandler(async (req, res) => {
     const { id: studentId } = getStudentValidator.parse(req.params);
+
+    const studentExists = await prisma.student.findUnique({ where: { id: studentId } });
+
+    if (!studentExists) {
+        throw new ApiError(HttpStatus.NOT_FOUND, "Student not found");
+    }
+
     await prisma.mark.deleteMany({
         where: { studentId: studentId },
     });
@@ -21,3 +29,4 @@ const deleteStudent: RequestHandler = asyncHandler(async (req, res) => {
 });
 
 export { deleteStudent };
+
